v0.01
I'm starting on just the website functionality. At this point in time,
it would be simpler to focus on the website first, over the mobile app,
because I'm not quite sure how the backend links together the site and
app using RESTful APIs.

** Note:  As of Febuary of 2017, I am not focused on the backend. This
site was inspired by Colt Steele's Web Developer Course and small
portions of code may be duplicated when related to the backend (RESTful
API and such). All other code will be custom fit by and for my own
ambitions.

With that said, let's get coding!